requirements.txt
astroid==1.6.3
bokeh==0.12.11
certifi==2017.11.5
chardet==3.0.4
colorama==0.3.9
cycler==0.10.0
decorator==4.1.2
diff-match-patch==20121119
Django==2.0.2
django-bootstrap4==0.0.5
django-filter==1.1.0
django-import-export==1.0.0
djangorestframework==3.8.2
et-xmlfile==1.0.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
isort==4.3.4
jdcal==1.3
jedi==0.11.0
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mysqlclient==1.3.12
nbformat==4.4.0
numpy==1.13.3
odfpy==1.3.6
openpyxl==2.5.2
panda==0.3.1
pandas==0.21.0
parso==0.1.0
pickleshare==0.7.4
plotly==2.2.2
prompt-toolkit==1.0.15
Pygments==2.2.0
pylint==1.8.4
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
tablib==0.12.1
tornado==4.5.2
traitlets==4.3.2
unicodecsv==0.14.1
urllib3==1.22
wcwidth==0.1.7
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
